Step_04_hello_agent/global-level/pyproject.toml
[project]
name = "global-level"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.1",
]
